refactor: Update import paths and rename index file
